technical-solution/ad-browser/agent/requirements.txt
# Ad Browser Agent 依赖清单
# 
# 安装方式（使用 UV）：
#   uv pip install -r requirements.txt
#
# 或使用传统 pip：
#   pip install -r requirements.txt

# 核心依赖
pydantic>=2.0.0          # 数据验证和模型定义
pydantic-settings>=2.0.0 # 配置管理（基于 pydantic）
python-dotenv>=1.0.0     # 环境变量加载

# Browser-Use 集成（浏览器自动化）
browser-use>=1.0.0       # Browser-Use 框架（延迟导入，可选）

# WebSocket 服务器（阶段四需要）
websockets>=12.0         # WebSocket 服务器实现

# 测试依赖（可选，用于开发）
pytest>=7.4.0            # 测试框架
pytest-asyncio>=0.21.0   # 异步测试支持
